+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Tighten generated pragmatic email-format validation and modularize primitive and constraint renderers.
+
 All notable changes to this project will be documented in this file.
 
 ## [0.0.4] - 2026-09-12